- changed placement of message filters
