+++ src/components/sidebar/L/tabs/MessageList.tsx
@@ -1,38 +1,13 @@
 import DMCard from "@/components/common/sidebar_buttons/DMCard";
 import LoadingFriendList from "@/components/common/skeletons/LoadingFriendList";
-import { messageList } from "@/data";
 import useClerkQuery from "@/hooks/use-query";
 import { Friends } from "@/types";
-import { Plus } from "lucide-react";
 
 const MessageList = () => {
 	const { data, isLoading, error } = useClerkQuery<Friends[]>("recent-chat");
 
 	return (
 		<div className=" grid overflow-y-auto max-h-full scrollbar-hidden">
-			{/* {data.map((message) => {
-				const {
-					profileImg,
-					user,
-					online,
-					hasMessage,
-					messageCount,
-					id,
-					slug,
-				} = message;
-				return (
-					<DMCard
-						key={id}
-						dmId={id}
-						slug={slug}
-						user={user}
-						profileImg={profileImg}
-						online={online}
-						hasMessage={hasMessage}
-						messageCount={messageCount}
-					/>
-				);
-			})} */}
 			{isLoading ? (
 				<LoadingFriendList />
 			) : error ? (
@@ -63,7 +38,7 @@ const MessageList = () => {
 					);
 				})
 			) : (
-				<div className="text-center">
+				<div className="text-center pt-5">
 					No recent conversations <br />
 					<p className="text-xs flex gap-1 justify-center pt-1">
 						Start a conversation with your friends
